% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AAPL}
\alias{AAPL}
\title{AAPL time series of daily OHLCVA bars}
\format{
An xts object with 501 rows and 6 variables:
\describe{
  \item{Open}{Daily Open prices}
  \item{High}{Daily High prices}
  \item{Low}{Daily Low prices}
  \item{Close}{Daily Close prices}
  \item{Volume}{Aggregate Daily volume traded}
  \item{Adjusted}{Adjusted prices}
}
}
\usage{
data(AAPL)
}
\description{
A dataset containing the daily OHLCVA values of AAPL from 2017-01-03 to
2018-12-28. The code to reproduce the dataset from yahoo is:
}
\details{
getSymbols(symbol,from=start_date, to = end_date, auto.assign = T, index.class = "POSIXct", src = 'yahoo')
for(i in symbols) assign(i, adjustOHLC(get(i),use.Adjusted=TRUE))

We have chosen to fix the dataset, since we have no control over changes at
the source, which would break our tests.
}
\keyword{datasets}
