\name{add.paramLookupTable}
\alias{add.paramLookupTable}
\title{add parameters to strategy objects: ALPHA CODE USE WITH CARE...}
\usage{add.paramLookupTable(strategy, type, name, paramTable)}
\description{add parameters to strategy objects: ALPHA CODE USE WITH CARE}
\arguments{\item{strategy}{strategy}
\item{type}{type}
\item{name}{name}
\item{paramTable}{paramTable}}
