% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{apply.paramset.signal.analysis}
\alias{apply.paramset.signal.analysis}
\title{Signal Analysis With Parmeter Optimization}
\usage{
apply.paramset.signal.analysis(strategy.st, paramset.label, portfolio.st,
  sigcol, sigval, on, forward.days, cum.sum = TRUE, include.day.of.signal,
  obj.fun, decreasing, mktdata = NULL, verbose = TRUE)
}
\arguments{
\item{strategy.st}{an object of type 'strategy' to add the indicator to}

\item{paramset.label}{a label uniquely identifying the paramset within the strategy}

\item{portfolio.st}{text name of the portfolio to associate the order book with}

\item{sigcol}{column name to check for signal}

\item{sigval}{signal value to match against}

\item{on}{the periods endpoints to find as a character string}

\item{forward.days}{days to exit post signal}

\item{cum.sum}{whether to use cumsum on price changes}

\item{include.day.of.signal}{whether to include the day of signal generation}

\item{obj.fun}{objective function for determining goodness of each paramset}

\item{decreasing}{how to sort the obj.fun output values}

\item{mktdata}{market data}

\item{verbose}{whether to output processing messages}
}
\value{
A list of objects that contains the results
\describe{
   \item{signals}{List of signals named by parameter combination. Each object within the list is a XTS object with columns being assets and rows being signals.}
   \item{sigret.by.paramset}{List of list. Sorted by parameter combination and then by asset. Contains matrix of next day returns; columns are period (N day look ahead) and rows are signals}
   \item{sigret.by.asset}{List of list. Sorted by Asset and then by parameter combinations. Contains same matrix as sigret.by.paramset}
   \item{signal.stats}{List of signal statistics as calculated via custom user defined fitness function. Sorted by assets.}
}
}
\description{
This function allows user to analyze signals independently from rules.
}
\details{
This function is similar to \code{applyParameter} but includes additionally \code{applySignal} wrapped within it. To use it, the user 
will need to initially define the distribution and constraints by using \code{add.distribution} and \code{add.distribution.constraint}.

More generally, the function is designed to generate signals across various parameter combinations defined within the distribution and
constraints. Afterwards it extract the next N day price changes for further anlaysis.

The parameter \code{on} allows user to calculate post signal returns in different frequencies. Different signals have different timeframes effectiveness. 
\code{forward.days} lets the user select the number of post signal days to analyze results.

To determine the goodness of a parameter set relative to all others, we let the user specify their own objective function (via parameter \code{obj.func})
that will be used to calculate an objective value representing each parameter. These will then be sorted from best to worst. The
objective function will take in a single matrix \code{ret.mat} of price changes. Each row represents an individual signal
while each column represents periods/post signal.

See demo 'signalAnalysisExample1.R' & 'signalAnalysisExample2.R'
}
\author{
Michael Guan
}
\seealso{
\code{\link{add.distribution}}
\code{\link{add.distribution.constraint}}
}

