% Generated by roxygen2 (4.0.2): do not edit by hand
\name{applyIndicatorSignals}
\alias{applyIndicatorSignals}
\title{Calculate Indicators and Signals for a Strategy}
\usage{
applyIndicatorSignals(strategy, portfolio, mktdata, sigcol, ...)
}
\arguments{
\item{strategy}{an object (or the name of an object) type 'strategy' to add the indicator to}

\item{portfolio}{text name of the portfolio to associate the order book with}

\item{mktdata}{market data}

\item{sigcol}{String name of signal to use for analysis}
}
\value{
\code{xts} object with columns representing signals for each symbol
}
\description{
Calculate signals given indicators and signal definitions.
}
\author{
Michael Guan
}
\seealso{
\code{\link{apply.paramset.signal.analysis}}
\code{\link{applyIndicators}}
\code{\link{applySignals}}
}

