\name{applyParameter}
\alias{applyParameter}
\title{Test different parameter sets on a strategy.}
\usage{
  applyParameter(strategy, portfolios, parameterPool,
  parameterConstrains, method, sampleSize)
}
\arguments{
  \item{strategy}{name of the strategy to apply paramters
  to.}

  \item{portfolios}{name of the portfolio to apply to.}

  \item{parameterPool}{a paramter distribution object that
  created by setParameterDistribution function, which
  includes all the parameter legal values and
  distribution/weights.}

  \item{parameterConstrains}{the object created by
  setParameterConstraint function that specifies the
  constrains between each parameters,}

  \item{method}{takes string 'expand' or 'random', specify
  how to generate samples of parameters. 'expand' will do
  all possible combinations of the parameter sets,}

  \item{sampleSize}{used when method=='random', specify how
  many parameter sets to generate and run test of.}
}
\description{
  Given a parameter distribution object generated by
  setParameterDistribution function, generate parameter
  sets and test each set on specified strategy.
}
\details{
  In the returned Object$: eachRun contains the details of
  test run with each parameter set. paramTable is parameter
  samples used, in a table for print. paramConstrainTable
  is the constraints apply to the parameters, in a table
  for print. parameterDistribution is the parameter
  distribution passed in as argument. parameterConstrains
  is the constraints apply to the parameters, passed in as
  argument.
}
\author{
  Yu Chen
}

