% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{applySignals}
\alias{applySignals}
\title{apply the signals in the strategy to arbitrary market data}
\usage{
applySignals(strategy, mktdata, indicators = NULL, parameters = NULL, ...)
}
\arguments{
\item{strategy}{an object of type 'strategy' to add the signal to}

\item{mktdata}{an xts object containing market data.  depending on signals, may need to be in OHLCV or BBO formats}

\item{indicators}{if indicator output is not contained in the mktdata object, it may be passed separately as an xts object or a list.}

\item{parameters}{named list of parameters to be applied during evaluation of the strategy}

\item{...}{any other passthru parameters}
}
\description{
This funcion is called internally by \code{\link{applyStrategy}} in normal 
operation, but it is also useful for nanual testing during development.
 
If you are using this function to test your strategy, note that the 'mktdata' 
argument should likely contain the output of \code{\link{applyIndicators}}.
}

