% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyStrategy.rebalancing.R
\name{applyStrategy.rebalancing}
\alias{applyStrategy.rebalancing}
\title{apply the strategy to arbitrary market data, with periodic rebalancing}
\usage{
applyStrategy.rebalancing(strategy, portfolios, mktdata = NULL,
  parameters = NULL, ..., verbose = TRUE, symbols = NULL,
  initStrat = FALSE, updateStrat = FALSE)
}
\arguments{
\item{strategy}{an object of type 'strategy' or the name of a stored strategy to apply}

\item{portfolios}{a list of portfolios to apply the strategy to}

\item{mktdata}{an xts object containing market data.  depending on indicators, may need to be in OHLCV or BBO formats, default NULL}

\item{parameters}{named list of parameters to be applied during evaluation of the strategy, default NULL}

\item{...}{any other passthru parameters}

\item{verbose}{if TRUE, return output list}

\item{symbols}{character vector identifying symbols to initialize a portfolio for, default NULL}

\item{initStrat}{whether to use (experimental) initialization code, default FALSE}

\item{updateStrat}{whether to use (experimental) wrapup code, default FALSE}
}
\description{
This function is the wrapper that holds together the execution of a strategy with rebalancing rules.
}
\seealso{
\code{\link{strategy}},  \code{\link{applyIndicators}}, 
 \code{\link{applySignals}}, \code{\link{applyRules}},
 \code{\link{initStrategy}}, \code{\link{applyStrategy}}
}

