\name{chart.forward.testing}
\alias{chart.forward.testing}
\title{Chart to analyse walk.forward() objective function}
\usage{
  chart.forward.testing(audit.filename)
}
\arguments{
  \item{audit.filename}{name of .audit environment file as
  produced by walk.forward()}
}
\description{
  Chart to analyse walk.forward() objective function
}

