% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramsets.R
\name{clone.orderbook}
\alias{clone.orderbook}
\title{clone a orderbook object, potentially stripping all history}
\usage{
clone.orderbook(
  orderbook.st,
  cloned.orderbook.st,
  strip.history = TRUE,
  src_envir = .strategy,
  target_envir = .strategy
)
}
\arguments{
\item{orderbook.st}{string identifying the source orderbook}

\item{cloned.orderbook.st}{string naming the target orderbook}

\item{strip.history}{boolean, default TRUE, whether to remove all orderbook history, keeping only the structure (e.g. assets)}

\item{src_envir}{environment to get the source orderbook from, default \code{.strategy}}

\item{target_envir}{environment to put the cloned orderbook in, passed to \code{\link{put.orderbook}}}
}
\description{
creates a copy of a orderbook, stripping all history (orders etc)
}
