% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramsets.R
\name{clone.portfolio}
\alias{clone.portfolio}
\title{clone a portfolio object, potentially stripping all history}
\usage{
clone.portfolio(portfolio.st, cloned.portfolio.st, strip.history = TRUE,
  src_envir = .blotter, target_envir = .blotter)
}
\arguments{
\item{portfolio.st}{string identifying the source portfolio}

\item{cloned.portfolio.st}{string naming the target portfolio}

\item{strip.history}{boolean, default TRUE, whether to remove all transaction and portfolio history, keeping only the structure (e.g. assets)}

\item{src_envir}{environment to get the portfolio to be cloned from, default \code{.blotter}}

\item{target_envir}{environment to put the cloned portfolio in, passed to \code{\link[blotter]{put.portfolio}}}
}
\description{
creates a copy of a portfolio, stripping all history (transactions etc)
}
