\name{delete.paramset}
\alias{delete.paramset}
\title{Delete a paramset from a strategy}
\usage{
  delete.paramset(strategy, paramset.label)
}
\arguments{
  \item{strategy:}{the name of the strategy object}

  \item{paramset.label:}{a label uniquely identifying the
  paramset within the strategy}
}
\description{
  Delete a paramset from a strategy
}
\author{
  Jan Humme
}

