% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{distributional.boxplot}
\alias{distributional.boxplot}
\title{Visualization of Single Signal}
\usage{
distributional.boxplot(signal, x.val = seq(1, 50, 5), val = 10,
  ylim = c(-5, 5), xlim = c(0, 50), mai = c(1, 1, 0.3, 0.5), h = 0, ...)
}
\arguments{
\item{signal}{list of paramset forward looking price changes by asset}

\item{x.val}{he values of the explanatory variable where we want to see the distribution}

\item{val}{this parameter determines how the plotted distribution is shown, increase/decrease it if the distribution is not shown properly}

\item{ylim}{the y limits in the plot}

\item{xlim}{the x limits in the plot}

\item{mai}{A numerical vector of the form c(bottom, left, top, right) which gives the margin size specified in inches.}

\item{h}{the y-value(s) for horizontal line(s).}

\item{...}{any other passthru parameters}
}
\value{
plot
}
\description{
This function employs \code{plotSimpleGamlss} in package \code{gamlss.util}.
}
\author{
Michael Guan
}

