% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramsets.R
\name{expand.distributions}
\alias{expand.distributions}
\title{given a distributions object expand all options to a full factorial representation}
\usage{
expand.distributions(distributions)
}
\arguments{
\item{distributions}{distributions slot from a strtaegy object}
}
\value{
expanded grid of all distributions as a `data.frame` by distribution labels
}
\description{
given a distributions object expand all options to a full factorial representation
}
\seealso{
expand.grid
}
