% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Federov.design.R
\name{Federov.constraints}
\alias{Federov.constraints}
\title{test a single parameter set to determine if it passes constraints}
\usage{
Federov.constraints(param.vec, ...)
}
\arguments{
\item{param.vec}{a vector (named or unamed) containing a trial parameter set}

\item{...}{any other passthrough parameters}
}
\value{
boolean -  TRUE if param.vec passes constraints, FALSE otherwise
}
\description{
test a single parameter set to determine if it passes constraints
}
\seealso{
Federov.paramset MonteCarlo.paramset apply.constraints
}
