% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Federov.design.R
\name{Federov.paramset}
\alias{Federov.paramset}
\title{construct a full Federov design for a strategy parameter set}
\usage{
Federov.paramset(
  strategy,
  paramset.label,
  ...,
  printd = FALSE,
  returnlist = FALSE,
  approximate = TRUE,
  center = TRUE
)
}
\arguments{
\item{strategy}{an object of type `strategy` that contains a parameter set to construct a Federov design for}

\item{paramset.label}{label describing the paramset to use in the strategy object}

\item{...}{any other passthrough parameters}

\item{printd}{if TRUE, print the design summary}

\item{returnlist}{if TRUE, return the list object describing the Federov design, else return just the `param.combos`}

\item{approximate}{if FALSE, use an exact design, will be slower but more accurate than if TRUE}

\item{center}{if TRUE, the default, center the parameters around a center value, see Details}
}
\value{

}
\description{
construct a full Federov design for a strategy parameter set
}
\seealso{
optFederov
}
