% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.strategy}
\alias{get.strategy}
\alias{getStrategy}
\title{retrieve strategy from the container environment}
\usage{
get.strategy(x, envir = .strategy)
}
\arguments{
\item{x}{string name of object to be retrieved}

\item{envir}{the environment to retrieve the strategy object from, defaults to .strategy}
}
\description{
retrieve strategy from the container environment
}

