\name{getParams}
\alias{getParams}
\title{get parameterized arguments list out of the strategy environment}
\usage{
  getParams(strategy, symbol, type, name)
}
\arguments{
  \item{strategy}{}

  \item{symbol}{}

  \item{type}{}

  \item{name}{}
}
\description{
  get parameterized arguments list out of the strategy
  environment
}

