\name{getStrategy}
\alias{getStrategy}
\title{retrieve strategy from the container environment}
\usage{
  getStrategy(x, envir = .strategy)
}
\arguments{
  \item{x}{string name of object to be retrieved}

  \item{envir}{the environment to retrieve the strategy
  object from, defaults to .strategy}
}
\description{
  retrieve strategy from the container environment
}

