% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initSymbol}
\alias{initSymbol}
\title{Run standard and custom symbol initialization functions}
\usage{
initSymbol(strategy, symbol, ...)
}
\arguments{
\item{strategy}{an object (or the name of an object) of type 'strategy' to add the init function definition to}

\item{symbol}{symbol}

\item{\dots}{any other passthrough parameters}
}
\description{
\code{initSymbol} will load a symbol and run user-defined functions to pre-process the symbol's data
before constructing indicators.
}
\details{
The custom initialization must be defined as named list containing
\describe{
  \item{name}{function name}
  \item{argument}{list of arguments}
  \item{enabled}{TRUE or FALSE}
}
and included as the slot \code{init_symbol} of the strategy object.
}

