% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramsets.R
\name{install.param.combo}
\alias{install.param.combo}
\title{insert a specific parameter combo into a strategy object}
\usage{
install.param.combo(strategy, param.combo, paramset.label)
}
\arguments{
\item{strategy}{strategy object}

\item{param.combo}{single parameter combination to be inserted}

\item{paramset.label}{label for the paramset to use to determine slot locations inside the strategy object}
}
\description{
In order to test \code{\link{applyStrategy}} with a specific parameter 
combination, it is necessary to insert those parameters into a copy of the 
strategy specification object.
}
\details{
This internal, non-exported function examines the paramset specification, 
and then uses that to insert the chosen parameters into a copy of the strategy
object.  It will search the strategy onject, component by component, and 
attempt to locate the components named in the paramset, so that the individual 
parameter values may be changed.
}
