% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{is.strategy}
\alias{is.strategy}
\title{test to see if object is of type 'strategy'}
\usage{
is.strategy(x)
}
\arguments{
\item{x}{object to be tested}
}
\description{
test to see if object is of type 'strategy'
}

