% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantstrat-package.R
\docType{data}
\name{luxoraudit}
\alias{luxoraudit}
\title{sample audit environment output from running luxor demo in file data/luxor.wfa.ples.RData}
\description{
sample audit environment output from running luxor demo in file data/luxor.wfa.ples.RData
}
\keyword{data}
