% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osFUNs.R
\name{osFixedPercent}
\alias{osFixedPercent}
\title{Order sizing function for fixing the position limit in percentage}
\usage{
osFixedPercent(data, timestamp, orderqty, portfolio, tradesize,
  leverage = 1, verbose = false, ...)
}
\arguments{
\item{data}{an xts object containing market data. depending on rules, may need to be in OHLCV or BBO formats, and may include indicator and signal information}

\item{timestamp}{timestamp coercible to POSIXct that will be the time the order will be inserted on}

\item{orderqty}{numeric quantity of the desired order, modified by osFUN}

\item{portfolio}{text name of the portfolio to get the equity}

\item{tradesize}{position limit in percentage, e.g. 0.3 for 30\%}

\item{leverage}{the leverage size, relevant for margin trading and derivatives market}

\item{verbose}{set it to true if you want to print logs on the console}

\item{...}{any other passthru parameters}
}
\description{
Allows to fix the position limit in percentage of the current portfolio equity.
}
\details{
Slows down the backtesting process as it updates the portfolio at each timestamp and not at the end of the timeseries.
}
\seealso{
\code{\link{addPosLimit}}, \code{\link{getPosLimit}}, \code{\link{osMaxPos}}
}
\author{
Valery Barmin \href{https://github.com/randomsuffer}{@randomsuffer}
}
