\name{osNoOp}
\alias{osNoOp}
\title{default order sizing function...}
\usage{osNoOp(orderqty, ...)}
\description{default order sizing function}
\details{default function performs no operation (NoOp), returns orderqty}
\arguments{\item{orderqty}{numeric quantity of the desired order, modified by osFUN}
\item{...}{any other passthru parameters}}
