\name{paramLookup}
\alias{paramLookup}
\title{add parameters to strategy objects: ALPHA CODE USE WITH CARE...}
\usage{paramLookup(strategy, symbol, type, name, parameter, ...)}
\description{add parameters to strategy objects: ALPHA CODE USE WITH CARE}
\arguments{\item{strategy}{strategy}
\item{symbol}{symbol}
\item{type}{type}
\item{name}{name}
\item{parameter}{parameter}
\item{...}{dots}}
