% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.signal.path}
\alias{plot.signal.path}
\title{Visualization of Signal Path}
\usage{
\method{plot}{signal.path}(data, main = "Cumulative Return Paths")
}
\arguments{
\item{data}{signal data}

\item{main}{plot title}
}
\value{
plot
}
\description{
This function creates a rChart - hplot that is interactive. It displays
}
\examples{
\dontrun{
# signalAnalysisExample1.R
plot.signal.path(results$sigret.by.asset$RTH$paramset.1.5[1:10,])
}
}
\author{
Michael Guan
}

