% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantstrat-package.R
\docType{data}
\name{portfolio.luxor}
\alias{portfolio.luxor}
\title{sample portfolio output from running luxor demo in file data/luxor-p066.RData}
\description{
sample portfolio output from running luxor demo in file data/luxor-p066.RData
}
\keyword{data}
