% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{post.signal.returns}
\alias{post.signal.returns}
\title{Generate Post Signal Returns}
\usage{
post.signal.returns(signals, sigval, on = NULL, forward.days,
  cum.sum = TRUE, include.day.of.signal = FALSE, mktdata = NULL)
}
\arguments{
\item{signals}{xts object with signals, one column}

\item{sigval}{signal value to match against}

\item{on}{the periods endpoints to find as a character string}

\item{forward.days}{number of days to look forward after signal (days to exit post signal)}

\item{cum.sum}{\code{TRUE},\code{FALSE}; cumulative sum of price changes}

\item{include.day.of.signal}{whether to analyze the return on signal day}

\item{mktdata}{market data}
}
\value{
\code{matrix} of post signal price changes; rows = nth signal, column = nth period since signal
}
\description{
This function collects and aggregates post signal price changes for N days forward.
}
\author{
Michael Guan
}
\seealso{
\code{\link{apply.paramset.signal.analysis}}
}

