% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degfreedom.R
\name{print.dof}
\alias{print.dof}
\title{print method for strategy degrees of freedom object}
\usage{
\method{print}{dof}(x, ...)
}
\arguments{
\item{x}{an object of type \code{dof} to be printed}

\item{...}{any other passthough parameters}
}
\description{
print method for strategy degrees of freedom object
}
\seealso{
\code{\link{degrees.of.freedom}}
}
