% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haircut.Sharpe.R
\name{print.haircutSR}
\alias{print.haircutSR}
\title{print method for Harvey and Liu Haircut Sharpe Ratio}
\usage{
\method{print}{haircutSR}(x, ...)
}
\arguments{
\item{x}{an object of type \code{haircutSR} to be printed}

\item{...}{any other passthough parameters}
}
\description{
print method for Harvey and Liu Haircut Sharpe Ratio
}
\seealso{
\code{\link{SharpeRatio.haircut}}
}
