% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profit.hurdle.R
\name{print.profitHurdle}
\alias{print.profitHurdle}
\title{print method for Harvey and Liu Haircut Sharpe Ratio}
\usage{
\method{print}{profitHurdle}(x, ...)
}
\arguments{
\item{x}{an object of type \code{profitHurdle} to be printed}

\item{...}{any other passthough parameters}
}
\description{
print method for Harvey and Liu Haircut Sharpe Ratio
}
\seealso{
\code{\link{profit.hurdle}}
}
