\name{quantstrat-package}
\alias{quantstrat-package}
\alias{quantstrat}
\docType{package}
\title{
Quantitative Strategy Model Framework
}
\description{
Transaction-oriented infrastructure for constructing trading systems and simulation.  Provides support for multi-asset class and multi-currency portfolios for backtesting and other financial research.  Still in heavy development.
}
\details{
}
\author{
Peter Carl
Brian G. Peterson

Maintainer: Brian G. Peterson <brian@braverock.com>
}
\keyword{ package }
\seealso{
\code{\link[quantmod:quantmod-package]{quantmod}}, 
\code{\link[blotter:blotter-package]{blotter}}, 
\code{\link[FinancialInstrument:FinancialInstrument-package]{FinancialInstrument}},
\code{\link[blotter:blotter-package]{blotter}}, 
\code{\link[PerformanceAnalytics:PerformanceAnalytics-package]{PerformanceAnalytics}}
}
\examples{
}
