% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance.rules.R
\name{rulePctEquity}
\alias{rulePctEquity}
\title{rule to base trade size on a percentage of available equity.}
\usage{
rulePctEquity(trade.percent = 0.02, ..., longlevels = 1, shortlevels = 1,
  digits = NULL, refprice = NULL, portfolio, symbol, timestamp)
}
\arguments{
\item{trade.percent}{max percentage of equity to allow the strategy to trade in this symbol}

\item{longlevels}{numeric number of levels}

\item{shortlevels}{numeric number of short levels, default longlevels}

\item{digits}{if not NULL(the default), will call \code{\link{round}} with specified number of digits}

\item{refprice}{if not NULL(the default), will divide the calculated tra}

\item{portfolio}{text name of the portfolio to place orders in, typically set automatically}

\item{symbol}{identifier of the instrument to cancel orders for, typically set automatically}

\item{timestamp}{timestamp coercible to POSIXct that will be the time the order will be inserted on, typically set automatically}

\item{\dots}{any other passthrough parameters}
}
\description{
This rule works with \code{\link{applyStrategy.rebalancing}} to set the
maximum trade size by calling \code{\link{addPosLimit}}.
}
\details{
To use it, you need to specify it as (part of) a rule of type 'rebalance'.
note that  \code{\link{applyStrategy.rebalancing}} will expect a 
'rebalance_on' argument to be included in the \code{arguments=list(...)} 
of the rule definition.
}
\examples{
# example rule definition
\dontrun{
add.rule(strategy.name, 'rulePctEquity',
        arguments=list(rebalance_on='months',
                       trade.percent=.02,
                       refprice=quote(last(getPrice(mktdata)[paste('::',curIndex,sep='')])[,1]),
                       digits=0
        ),
        type='rebalance',
        label='rebalance')
}
}
\seealso{
\code{\link{osMaxPos}} , 
\code{\link{applyStrategy.rebalancing}}, 
\code{\link{addPosLimit}}, 
\code{\link{add.rule}}
}

