% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haircut.Sharpe.R
\name{sample_random_multests}
\alias{sample_random_multests}
\title{Generate empirical p-value distributions}
\usage{
sample_random_multests(rho, m_tot, p_0, lambda, M_simu)
}
\arguments{
\item{rho}{average correlation among returns}

\item{m_tot}{total number of trials}

\item{p_0}{probability for a random factor to have a zero mean}

\item{lambda}{average of monthly mean returns for true strategies}

\item{M_simu}{number of simulations to perform}
}
\description{
Generate empirical p-value distributions
}
