% Generated by roxygen2 (4.0.2): do not edit by hand
\name{save.strategy}
\alias{save.strategy}
\title{save a strategy object from memory onto disk}
\usage{
save.strategy(strategy.name)
}
\arguments{
\item{strategy.name}{a string specifying the name of the strategy object; will also be used to create a file name}
}
\description{
save a strategy object from memory onto disk
}

