% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflated.Sharpe.R
\name{deflatedSharpe}
\alias{deflatedSharpe}
\alias{SharpeRatio.deflated}
\alias{.deflatedSharpe}
\title{Calculate a Deflated Sharpe Ratio using number of trials and portfolio moments}
\usage{
deflatedSharpe(portfolios, ..., strategy = NULL, trials = NULL,
  audit = NULL, env = .GlobalEnv)

.deflatedSharpe(sharpe, nTrials, varTrials, skew, kurt, numPeriods,
  periodsInYear = 252)
}
\arguments{
\item{portfolios}{string name of portfolio, or optionally a vector of portfolios, see DETAILS}

\item{...}{any other passtrhrough parameters}

\item{strategy}{optional strategy specification that would contain more information on the process, default NULL}

\item{trials}{optional number of trials,default NULL}

\item{audit}{optional audit environment containing the results of parameter optimization or walk forward, default NULL}

\item{env}{optional environment to find market data in, if required.}

\item{sharpe}{candidate (annualized) Sharpe Ratio}

\item{nTrials}{numeric number or trials}

\item{varTrials}{variance of Sharpe ratios of the trials}

\item{skew}{skewness of the candidate}

\item{kurt}{non-excess kurtosis}

\item{numPeriods}{total periods in the backtest}

\item{periodsInYear}{number of periods in a year, default 252 (daily)}
}
\value{
a \code{data.frame} containing:

\itemize{
  \item{original observed Sharpe ratio}
  \item{deflated Sharpe ratio}
  \item{p-value of the deflated Sharpe ratio}
  \item{number of trials used for adjustment}
}

this object may change in the future, and may be classed so that we can include more information
}
\description{
Per Bailey and Lopex de Prado (2014), construct a Deflated Sharpe Ratio and 
associated p-value based on an observed Sharpe ratio and information drawn
from a series of trials (e.g. parameter optimization or other strategies tried
before the candidate strategy)

internal implementation of Deflated Sharpe
}
\references{
Bailey, David H, and Marcos López de Prado. 2014. “The Deflated Sharpe Ratio: 
Correcting for Selection Bias, Backtest Overfitting and Non-Normality.” 
Journal of Portfolio Management 40 (5): 94-107. 
http://www.davidhbailey.com/dhbpapers/deflated-sharpe.pdf

https://quantstrattrader.wordpress.com/2015/09/24/
}
\author{
Brian G. Peterson

Ilya Kipnis, Brian G. Peterson
}
