\name{sigComparison}
\alias{sigComparison}
\title{signal function for comparisons...}
\usage{sigComparison(label, data, columns, relationship=c("gt", "lt", "eq", "gte", "lte"))}
\description{signal function for comparisons}
\arguments{\item{label}{text label to apply to the output}
\item{data}{data to apply comparison to}
\item{columns}{named columns to apply comparison to}
\item{relationship}{one of c("gt","lt","eq","gte","lte") or reasonable alternatives}}
