% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{signal.obj.slope}
\alias{signal.obj.slope}
\title{Signal Objective Function}
\usage{
signal.obj.slope(x)
}
\arguments{
\item{x}{Return matrix, each row represents price deltas for a single signal, each column represents periods}
}
\value{
Single Objective Value
}
\description{
Simple example of objective function that can can be used to measure the effectiveness of various parameter combinations.
}
\details{
Calculates the slope of the Cumulative Equity line. Higher the better.

It is important to note that all objective functions are called within \code{signal.generate.statistics}. What gets
passed in to it as parameter is an matrix of post signal price changes or cumulative equity for each signal. The matrix
is of N-by-M dimensions. N being the row representing the Nth signal. M being the column representing Mth period.
}
\author{
Michael Guan
}
\seealso{
\code{\link{apply.paramset.signal.analysis}}
}

