% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{signal.path.plot}
\alias{signal.path.plot}
\title{Visualization of Signal Path}
\usage{
signal.path.plot(data, main = "Cumulative Return Paths")
}
\arguments{
\item{data}{signal data}

\item{main}{plot title}
}
\value{
plot
}
\description{
This function creates a rChart - hplot that is interactive. It displays
}
\examples{
\dontrun{
# signalAnalysisExample1.R
signal.plot.path(results$sigret.by.asset$RTH$paramset.1.5[1:10,])
}
}
\author{
Michael Guan
}

