% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{signal.plot}
\alias{signal.plot}
\title{Visualization of Signal Across Lookback}
\usage{
signal.plot(signals, rows = NULL, columns = NULL, mai = c(0.1, 0.4, 0.2,
  0.1), mgp = c(1, 1, 0), xlab = "", ylab = "", cex.main = 0.6,
  xaxt = "n", cex.axis = 0.5, h = 0, hlinecol = "red", ...)
}
\arguments{
\item{signals}{list of paramset forward looking price changes by asset}

\item{rows}{number of rows for plot}

\item{columns}{number of columns for plot}

\item{mai}{A numerical vector of the form c(bottom, left, top, right) which gives the margin size specified in inches.}

\item{mgp}{The margin line (in mex units) for the axis title, axis labels and axis line. Note that mgp[1] affects title whereas mgp[2:3] affect axis. The default is c(3, 1, 0).}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{cex.main}{The magnification to be used for main titles relative to the current setting of cex.}

\item{xaxt}{A character which specifies the x axis type. Specifying "n" suppresses plotting of the axis. The standard value is "s": for compatibility with S values "l" and "t" are accepted but are equivalent to "s": any value other than "n" implies plotting.}

\item{cex.axis}{The magnification to be used for axis annotation relative to the current setting of cex.}

\item{h}{the y-value(s) for horizontal line(s).}

\item{hlinecol}{A specification for the default plotting color. See section \sQuote{Color Specification}.}

\item{...}{any other passthru parameters}
}
\description{
This function takes a list of  matrix of post signal price changes and 
plots boxplots. Note function plots whatever is given to it therefore when 
there are lots paramsets, it is best to plot a smaller portion so the 
information can be displayed clearly.
}
\author{
Michael Guan
}

