\name{sigThreshold}
\alias{sigThreshold}
\title{signal function for threshold signal...}
\usage{sigThreshold(label, data, column, threshold=0, relationship=c("gt", "lt", "eq", "gte", "lte"))}
\description{signal function for threshold signal}
\arguments{\item{label}{text label to apply to the output}
\item{data}{data to apply comparison to}
\item{column}{named column to apply comparison to}
\item{relationship}{one of c("gt","lt","eq","gte","lte") or reasonable alternatives}}
