% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantstrat-package.R
\docType{data}
\name{stats}
\alias{stats}
\title{sample tradeStats output from running luxor demo in file data/luxor.parameters.1-10.30-55.RData}
\description{
sample tradeStats output from running luxor demo in file data/luxor.parameters.1-10.30-55.RData
}
\keyword{data}
