\name{updateOrderMatrix}
\alias{updateOrderMatrix}
\title{insert a block of updated orders...}
\usage{updateOrderMatrix(portfolio, symbol, updatedorders)}
\description{insert a block of updated orders}
\arguments{\item{portfolio}{text name of the portfolio to associate the order book with}
\item{symbol}{identfier of the instrument to find orders for.  The name of any associated price objects (xts prices, usually OHLC) should match these}
\item{updatedorders}{time series containing updated orders}}
