\name{add.parameter}
\alias{add.parameter}
\title{add parameters to strategy objects: ALPHA CODE USE WITH CARE...}
\usage{add.parameter(strategy, type=c("indicator", "signal"), add.to.name,
    method=c("lookup", "lookup.range", "calc"), arguments, label, ...,
    store=FALSE)}
\description{add parameters to strategy objects: ALPHA CODE USE WITH CARE}
\arguments{\item{strategy}{}
\item{type}{}
\item{add.to.name}{}
\item{method}{}
\item{arguments}{}
\item{label}{}
\item{...}{}
\item{store}{}}
