\name{add.paramLookupTable}
\alias{add.paramLookupTable}
\title{add parameters to strategy objects: ALPHA CODE USE WITH CARE...}
\usage{add.paramLookupTable(strategy, type, name, paramTable)}
\description{add parameters to strategy objects: ALPHA CODE USE WITH CARE}
\arguments{\item{strategy}{}
\item{type}{}
\item{name}{}
\item{paramTable}{}}
