% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{add.signal}
\alias{add.signal}
\title{add a signal to a strategy}
\usage{
add.signal(strategy, name, arguments, parameters = NULL, label = NULL, ...,
  enabled = TRUE, indexnum = NULL, store = FALSE)
}
\arguments{
\item{strategy}{an object (or the name of an object) of type 'strategy' to add the signal to}

\item{name}{name of the signal, must correspond to an R function}

\item{arguments}{named list of default arguments to be passed to an signal function when executed}

\item{parameters}{vector of strings naming parameters to be saved for apply-time definition,default NULL, only needed if you need special names to avoid argument collision}

\item{label}{arbitrary text label for signal output, default NULL}

\item{...}{any other passthru parameters}

\item{enabled}{TRUE/FALSE whether the signal is enabled for use in applying the strategy, default TRUE}

\item{indexnum}{if you are updating a specific signal, the index number in the $signals list to update}

\item{store}{TRUE/FALSE whether to store the strategy in the .strategy environment, or return it.  default FALSE}
}
\value{
if \code{strategy} was the name of a strategy, the name. It it was a strategy, the updated strategy.
}
\description{
This adds a signal definition to a strategy object.
}
\details{
Signals denote times at which the strategy \emph{may} want to 
take action.  Common signals types from the literature include 
crossovers, thresholds, or other interactions between your \code{mktdata}
and your indicators.

if \code{label} is not supplied,  NULL default will be converted to '<name>.sig'
if the signal function returns one named column, we use that, and ignore the label.  
If the signal function returns multiple columns, the label will be 
\code{\link{paste}}'d to either the returned column names or the 
respective column number.
}
\seealso{
\code{\link{applySignals}}
\code{\link{add.indicator}}
\code{link{add.rule}}
\code{\link{sigComparison}}
\code{\link{sigCrossover}}
\code{\link{sigFormula}}
\code{\link{sigPeak}}
\code{\link{sigThreshold}}
}

