\name{applyStrategy}
\alias{applyStrategy}
\title{apply the strategy to arbitrary market data...}
\usage{applyStrategy(strategy, mktdata, ...)}
\description{apply the strategy to arbitrary market data}
\arguments{\item{strategy}{an object of type 'strategy' to add the indicator to}
\item{mktdata}{an xts object containing market data.  depending on indicators, may need to be in OHLCV or BBO formats}
\item{...}{any other passthru parameters}}
