% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.forward.R
\name{chart.forward}
\alias{chart.forward}
\title{Chart to analyse walk.forward() objective function}
\usage{
chart.forward(audit.filename, portfolio.st = NULL)
}
\arguments{
\item{audit.filename}{name of .audit environment file as produced by \code{\link{walk.forward.}}
Filename will often match pattern [audit.prefix].results.RData. Alternately, an
audit environment provided by the output of \code{\link{walk.forward}}}

\item{portfolio.st}{string defining which portfolio should be used for out of sample, default NULL}
}
\description{
The \code{\link{walk.forward}} function creates an audit environment,  potentially written 
out as a results file, which contains the out of sample results of the (chosen)
parameter set.  This function parses that file  or environment and generates a chart to 
compare the performance of all parameter sets against the optimal one.
}
\details{
\code{\link{chart.forward.training}} uses the audit environment for a single 
in-sample training period to draw a chart for that in-sample data.

Note that parameter \code{audit filename} may also be an audit environment 
which is already loaded in \R, for ease of development and debugging.  Little
checking is done to ensure the correct structure of this environment, so passing
an unsuitable environment (such as the .blotter environment) will result in
errors.
}
\seealso{
\code{\link{walk.forward}}, \code{\link{chart.forward.training}}
}
