% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.forward.R
\name{chart.forward}
\alias{chart.forward}
\title{Chart to analyse walk.forward() objective function}
\usage{
chart.forward(audit.filename)
}
\arguments{
\item{audit.filename}{name of .audit environment file as produced by walk.forward().
Filename will match pattern [audit.prefix].results.RData.}
}
\description{
Chart to analyse walk.forward() objective function
}

