% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chart.forward.training}
\alias{chart.forward.training}
\title{Chart to analyse walk.forward() objective function}
\usage{
chart.forward.training(audit.filename)
}
\arguments{
\item{audit.filename}{name of .audit environment file as produced by walk.forward()}
}
\description{
Chart to analyse walk.forward() objective function
}

