% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.forward.training.R
\name{chart.forward.training}
\alias{chart.forward.training}
\title{Chart to analyse walk.forward() objective function}
\usage{
chart.forward.training(audit.filename)
}
\arguments{
\item{audit.filename}{name of .audit environment file as produced by walk.forward().
Filename will match pattern [audit.prefix].[symbol].[start timestamp].[end timestamp].RData.}
}
\description{
The \code{\link{walk.forward}} function creates an audit environment, written 
out as a results file, which contains the out of sample results of the (chosen)
parameter set. It also writes out in-sample audit environment files for each 
training period.
}
\details{
This function reads one of those training files and will construct a performance 
chart of the in-sample performance of all the parameter sets during the training
period, including the in-sample performance of the chosen parameter set.

Note that parameter \code{audit filename} may also be an audit environment 
which is already loaded in \R, for ease of development and debugging.  Little
checking is done to ensure the correct structure of this environment, so passing
an unsuitable environment (such as the .blotter environment) will result in
errors.
}
\seealso{
\code{\link{walk.forward}}, \code{\link{chart.forward}}
}
