% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.forward.training.R
\name{chart.forward.training}
\alias{chart.forward.training}
\title{Chart to analyse walk.forward() objective function}
\usage{
chart.forward.training(audit.filename)
}
\arguments{
\item{audit.filename}{name of .audit environment file as produced by walk.forward().
Filename will match pattern [audit.prefix].[symbol].[start timestamp].[end timestamp].RData.}
}
\description{
Chart to analyse walk.forward() objective function
}

