% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profit.hurdle.R
\name{.profitHurdle}
\alias{.profitHurdle}
\title{internal implementation of profit hurdle code, see \code{\link{profit.hurdle}}}
\usage{
.profitHurdle(sm_fre, num_tests, num_obs, alpha_sig, vol_annual, RHO)
}
\arguments{
\item{sm_fre}{}

\item{num_tests}{}

\item{num_obs}{}

\item{alpha_sig}{}

\item{vol_annual}{}

\item{RHO}{}
}
\description{
internal implementation of profit hurdle code, see \code{\link{profit.hurdle}}
}
