\name{getOrderBook}
\alias{getOrderBook}
\title{get the order book object...}
\usage{getOrderBook(portfolio)}
\description{get the order book object}
\details{I don't think this should be exported, but it is for now while we're in test mode.}
\arguments{\item{portfolio}{text name of the portfolio the order book is associated with}}
