\name{getParams}
\alias{getParams}
\title{get parameterized arguments list out of the strategy environment...}
\usage{getParams(strategy, symbol, type, name)}
\description{get parameterized arguments list out of the strategy environment}
\arguments{\item{strategy}{strategy}
\item{symbol}{symbol}
\item{type}{type}
\item{name}{name}}
