% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initStrategy}
\alias{initStrategy}
\title{run standard and custom strategy initialization functions}
\usage{
initStrategy(strategy, portfolio, symbols, parameters = NULL,
  get.Symbols = FALSE, init.Portf = TRUE, init.Acct = TRUE,
  init.Orders = TRUE, unique = TRUE, ...)
}
\arguments{
\item{strategy}{object of type \code{strategy} to initialize data/containers for}

\item{portfolio}{portfolio}

\item{symbols}{symbols}

\item{parameters}{named list of parameters to be applied during evaluation of the strategy, default NULL}

\item{get.Symbols}{TRUE/FALSE, default FALSE}

\item{init.Portf}{TRUE/FALSE, default TRUE}

\item{init.Acct}{TRUE/FALSE, default TRUE}

\item{init.Orders}{TRUE/FALSE, default TRUE}

\item{unique}{TRUE/FALSE, default TRUE}

\item{\dots}{any other passthrough parameters}
}
\description{
\code{initStrategy} will run a series of common initialization functions at the 
beginning of an \code{\link{applyStrategy}} call.
}
\details{
\describe{
     \item{get.Symbols}{if TRUE, will call \code{\link[quantmod]{getSymbols}} 
                         on all symbols included in the \code{symbols} vector}
     \item{init.Portf}{if TRUE, will call \code{\link[blotter]{initPortf}} 
                         to initialize the portfolio object}
     \item{init.Acct}{if TRUE, will call \code{\link[blotter]{initAcct}} 
                         to initialize the account object}
     \item{init.Orders}{if TRUE, will call \code{\link{initOrders}} 
                         to initialize the order book for this test}
     \item{unique}{not yet implemented, will force a unique portfolio and account name 
                         if the portfolio, account, or order book already exist}
}

If used in conjuction with \code{initBySymbol}, \code{get.Symbols} should be \code{FALSE}.
}
\author{
Garrett See, Brian Peterson
}
\seealso{
\code{\link{applyStrategy}}, \code{\link{add.init}},
}

