% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{load.strategy}
\alias{load.strategy}
\title{load a strategy object from disk into memory}
\usage{
load.strategy(strategy.name, file = NULL)
}
\arguments{
\item{strategy.name}{a string specifying the name of the strategy object; may also be used to create a file name}

\item{file.name}{string specifying the filename to use, if NULL, a default will be created}
}
\description{
load a strategy object from disk into memory
}
