% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{paramConstraint}
\alias{paramConstraint}
\title{Internal function used in applyParameter function for process constraints on relationship between two parameter values. (deprecated)}
\usage{
paramConstraint(label, data = mktdata, columns, relationship = c("gt", "lt",
  "eq", "gte", "lte", "op"))
}
\arguments{
\item{label}{text label to apply to the constraint}

\item{data}{data to apply comparison to}

\item{columns}{named columns to apply comparison to}

\item{relationship}{one of c("gt","lt","eq","gte","lte","op") or reasonable alternatives}
}
\description{
Basically is the same as sigComparison function in signal.R written by Brian, with minor change.
}
\details{
Currently, this function compares two columns.  
Patches to compare an arbitrary number of columns would be gladly accepted.

Comparison will be applied from the first to the second column in the \code{columns} vector.

Relationship 'op' means 'opposite' side.  Reasonable attempt will be made to match.
}

