\name{paramLookup}
\alias{paramLookup}
\title{add parameters to strategy objects: ALPHA CODE USE WITH CARE...}
\usage{paramLookup(strategy, symbol, type, name, parameter, ...)}
\description{add parameters to strategy objects: ALPHA CODE USE WITH CARE}
\arguments{\item{strategy}{}
\item{symbol}{}
\item{type}{}
\item{name}{}
\item{parameter}{}
\item{...}{}}
