% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{put.strategy}
\alias{put.strategy}
\title{put a strategy object in .strategy env}
\usage{
put.strategy(strategy, envir = .strategy)
}
\arguments{
\item{strategy}{object; name will be extracted as strategy$name}

\item{envir}{the environment to store the strategy in, defaults to .strategy}
}
\description{
put a strategy object in .strategy env
}
\seealso{
getStrategy
}

