\name{setParameterConstraint}
\alias{setParameterConstraint}
\title{Function to construct parameter constraint object.}
\usage{
  setParameterConstraint(paramConstraintObj = list(),
  constraintLabel, paramList, relationship)
}
\arguments{
  \item{constrainlabel}{give a label to the constraint.}

  \item{paramList}{the two name of the prameters as a list
  contains two strings.}

  \item{relationship}{relationship between the 1st
  parameter and 2nd one. ('gt' means 1st parameter > 2nd
  parameter)}
}
\description{
  Function to construct parameter constraint object.
}
\author{
  Yu Chen
}

