% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haircut.Sharpe.R
\name{haircutSharpe}
\alias{haircutSharpe}
\alias{.haircutSR}
\title{Haircut Sharpe Ratio to correct for number of trials and autocorrelation}
\usage{
haircutSharpe(portfolios, ..., strategy = NULL, trials = NULL,
  audit = NULL, env = .GlobalEnv)

.haircutSR(sm_fre, num_obs, SR, ind_an, ind_aut, rho, num_test, RHO)
}
\arguments{
\item{portfolios}{string name of portfolio, or optionally a vector of portfolios, see DETAILS}

\item{...}{any other passtrhrough parameters}

\item{strategy}{optional strategy specification that would contain more information on the process, default NULL}

\item{trials}{optional number of trials,default NULL}

\item{audit}{optional audit environment containing the results of parameter optimization or walk forward, default NULL}

\item{env}{optional environment to find market data in, if required.}

\item{sm_fre}{Sampling frequency; [1,2,3,4,5] = [Daily, Weekly, Monthly, Quarterly, Annual]}

\item{num_obs}{No. of observations in the frequency specified in the previous step}

\item{SR}{Sharpe ratio; either annualized or in the frequency specified in the previous step}

\item{ind_an}{Indicator; if annulized, 'ind_an' = 1; otherwise = 0}

\item{ind_aut}{Indicator; if adjusted for autocorrelations, 'ind_aut' = 0; otherwise = 1}

\item{rho}{Autocorrelation coefficient at the specified frequency}

\item{num_test}{'num_test': Number of tests allowed. e.g. Harvey, Liu and Zhu (2014) find 315 published equity risk factors}

\item{RHO}{Average correlation among contemporaneous strategy returns}
}
\value{
NONE yet, sorry!
}
\description{
Haircut Sharpe Ratio to correct for number of trials and autocorrelation

The non-exported \code{.haircutSR} function is the internal implementation of
the haircut Sharpe Ratio calculations. It is based off of the code  written 
by Jasen Mackie for his blog post referenced below. It calcualtes an equivalent
adjusted Sharpe Ratio after taking the number and autocorrelation of the
trials into account.
}
\references{
Harvey, Campbell R. and Yan Liu. 2015. Backtesting The Journal of Portfolio Management. 41:1 pp. 13-28.#' @importFrom TTR ROC

Mackie, Jasen. 2016. R-view: Backtesting - Harvey & Liu (2015). https://opensourcequant.wordpress.com/2016/11/17/r-view-backtesting-harvey-liu-2015/
}
\author{
Jasen Mackie, Brian G. Peterson
}
