% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{sigCrossover}
\alias{sigCrossover}
\title{generate a crossover signal}
\usage{
sigCrossover(label, data = mktdata, columns, relationship = c("gt", "lt",
  "eq", "gte", "lte"), offset1 = 0, offset2 = 0)
}
\arguments{
\item{label}{text label to apply to the output}

\item{data}{data to apply crossover to}

\item{columns}{named columns to apply crossover of the first against the second}

\item{relationship}{one of c("gt","lt","eq","gte","lte") or reasonable alternatives}

\item{offset1}{numeric offset to be added to the first column prior to comparison}

\item{offset2}{numeric offset to be added to the second column prior to comparison}
}
\description{
This will generate a crossover signal, which is a dimension-reduced version 
of a comparison signal \code{\link{sigComparison}}.
}
\details{
It will return TRUE on the period in which there is a crossover in the 
direction specified by \code{relationship}, and NA otherwise.

If you want all the information, use a comparison instead.
}

