% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals.R
\name{signal.generate.statistics}
\alias{signal.generate.statistics}
\title{Signal Objective Function Calculation}
\usage{
signal.generate.statistics(post.ret, obj.fun = NULL, decreasing = TRUE)
}
\arguments{
\item{post.ret}{\code{matrix} of parameter set of post signal price deltas}

\item{obj.fun}{custom objective function for measuring signal goodness}

\item{decreasing}{if true, the higher the objective value, the better}
}
\value{
objective function values
}
\description{
This function takes a list of matrix of post signal price changes and 
applies an user defined objective function on it.
}
\author{
Michael Guan
}
\seealso{
\code{\link{apply.paramset.signal.analysis}}
}

